[FEAT][UI/UX]로그인/회원가입(+회원약관동의) 디자인 수정, 반응형 작업
- tailwind css의 color var 대신 우리 프로젝트의 primary color로 수정
- 반응형 간단하게 조치
