+++ src/pages/account/LoginPage/Description.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 function Description() {
   return (
-    <div className='bg-blue-500 bg-opacity-50' style={{ flexBasis: '50%'}}>
+    <div className='lg:block hidden bg-blue-500 bg-opacity-50 w-[50%]'>
         <h1 style={{ color:'white'}}>Left</h1>
     </div>
   )
