+++ src/pages/account/SignupPage/termsOfUseData.json
@@ -1,0 +1,22 @@
+[
+    {
+      "목적": "LiQuest플랫폼 일반 회원가입 및 서비스의 이용, 유지, 종료",
+      "항목": "회원가입 신청인의 성명(국문 및 영문), 이메일(아이디), 비밀번호, 휴대전화번호, 성별, 생년월일, 연계정보(CI), 회사코드(기업고객에 한함)",
+      "보유기간": "LiQuest플랫폼 제공 서비스 이용기간 동안.분쟁대비를 위해 서비스 이용기간 종료 후 6개월까지.단, 분쟁지속시 해결시 까지 보유※ 예외법령에 특별한 규정이 있을 경우 관련 법령에 따라 보관"
+    },
+    {
+      "목적": "LiQuest 네트워크 서비스 제공, 이용관련 문의 불만처리, 고객민원처리 등",
+      "항목": "성명, 이메일, 사진, 소셜미디어 계정정보",
+      "보유기간": "개인 페이지 삭제시 또는 회원탈퇴시까지 보유"
+    },
+    {
+      "목적": "관련 법률에 따르거나 서비스 이용 또는 업무처리과정에서 생성되어 수집/이용",
+      "항목": "이용시간/이용기록, 접속로그, 이용 컨텐츠, 쿠키, 접속IP정보, 결제기록, 이용정지기록 등 서비스 이용정보",
+      "보유기간": "LiQuest플랫폼 제공 서비스 이용기간"
+    },
+    {
+      "목적": "LiQuest가 제공하는 이용자 맞춤형 서비스 제공",
+      "항목": "성명, 이메일(아이디), 휴대폰정보, 이용시간/이용기록, 접속로그, 이용 컨텐츠, 쿠키, 접속IP정보, 결제기록, 이용정지기록 등 서비스 이용정보",
+      "보유기간": "LiQuest플랫폼 제공 서비스 이용기간"
+    }
+]
